+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add oauth2 and OIDC authentication backends for SSO with Open edX
 - Add an icon to the 'login' button.
+- Add icons in course glimpse footer & next to organization name.
 
 ## [2.0.0-beta.8] - 2020-06-17